Fix dangling `yes` on macOS
On macOS and BSD ignored signals are ignored in child processes too.
This creates an issue together with the fact that BEAM ignores SIGPIPE
signal. If we run `yes | docker-compose` on these platforms, the `yes`
program won't get the SIGPIPE signal after `docker-compose` exits.
`yes` then continues in trying to write 'y' to the closed pipe and
does that ad infinitum while using 100% of 1 CPU core.

This commit replaces `yes` with `printf 'y\n%.0s' {1..1000}` which
prints 'y' only 1000 times which solves the issue as the process
will exit after a short time even when it doesn't receive the SIGPIPE.
Hopefully, 1000 times should be enough for all cases.